Hide list of schools on profile for LAs
We now hide the list of associated schools from a hiring staff user on a school group's
profile if the school group is a local authority. This is because all schools are
associated with the LA they are within, geographically. This means that all local authority
users will be able to change the profiles of all schools within their geographic area.

We are unsure of the policy implications for this, so, for now, have decided to take the
ability to do this away whilst we explore this issue. This PR only hides the list of
schools on a school group's profile page if that school group is an LA.

We should probably prevent LAs from being able to reach the edit page full stop (although
it's quite unlikely that they would visit it as they would need the direct URL). This will
be easier as a result of changes I will (hopefully) shortly introduce with a PR for
https://dfedigital.atlassian.net/browse/TEVA-4476.