Add organisation profile page & publisher preview
Previously we declared the resource route "schools". This has been changed to "organisations",
with the option path: "schools" so the URL doesn't clash with our organisation landing page.
Instead, the schools resource route is now nested within the organisations resource route.

If the organisation is a trust, we now allow users to click on a tab on the organisation's
profile to see a list of schools at that trust - an index page for all of the schools that
belong to the trust. For this reason, a schools resource route is now nested within the
organisations resource route. This schools resource route is restricted to only an index route
for this schools index page. A show route has been added to the `organisations` resource for
showing an organisation's profile and an index route has been added for the organisation search
results page.