Save vacancy before sending document event
Our team's performance analyst noticed a problem with supporting document created
events coming from the application forms controller. Many of the events contained
null values. After investigating this issue, I found it was caused by the order
in which things are done in the create action of the application forms controller.

Looking at the docs for ActiveStorage::Attached::One's attach method
revealed the following:

"If the record is persisted and unchanged, the attachment is saved to the database
immediately. Otherwise, it'll be saved to the DB when the record is next saved."

In our case, the form which contains the file upload field also contains another
field - the application email field. If I user were to attach an application form
and at the same time change the value for application_email, this record (the
vacancy) will be in a changed state, meaning the attached file will be saved when
the record is next saved.

As the record is saved in the create action (vacancy.update(form.params_to_save))
after the event is sent, the application form isn't actually attached at the point
where the event is sent, causing the null values.