Use paths instead of paths-ignore in workflow
A previous attempt (commit 7253545dc1db67d09c7b52e456459ccc3a8539e9) to fix a problem
with changes to long-form content being deployed was unsuccessful. After viewing some of
the documentation, it seems that paths must be used along with negative patterns (prefixed
with !) if we want to ignore changes to files in certain locations but add exceptions.