Revert jest to 29.4.3
For whatever reason, jest 29.5.0 is causing a few tests in app/assets/javascript/components/map/map.test.js
to fail. It seems that this TypeError (Cannot use 'in' operator to search for '_leaflet_id' in null) is raised
at teaching-vacancies/node_modules/leaflet/src/core/Util.js:55:21, so it could be that leaflet doesn't work
particularly well with the latest minor version of jest.