Avoid Redis&Sidekiq major V. Dependabot upgrades
We cannot upgrade Redis & Sidekiq major gem versions until we use a
Redis service instance at version 6.2 or greater.

Our current Redis service instance on GovUK PaaS is v5.0.6.

Dependabot is opening PRs trying to upgrade both gems to their latest
versions. And when the PR gets closed/blocked by a developer, Dependabot
overseeds/reopens a new PR with a further patch version increase.

This causes the developer visibility on "why we cannot upgrade these" to
get lost as soon as a new version triggers a PR.

To avoid the hassle of managing this on each version, and to mitigate
the risk of merging a PR that would break our builds, we are ignoring any
major version upgrades for both gems.

Opted for an explicit configuration file definitions instead of simply
calling a "@dependabot ignore xxx" in the PR comment, as this allows us
to have an explicit configuration in our code with comments reminding us
about the need to upgrade these gems in the future once the blocker is
removed.

A "@dependabot ignore xxx" call would put us at risk of completely
forgetting to upgrade the affected gems.