Increase production web app instances to 8
We have been receiving spikes of traffic daily that caused our app
instances to crash due to high CPU usage.

The mitigation has been to manually temporally scale the application web
instances from 4 to 8.

As this is becoming a daily issue, we are permanently setting the number
of instances to 8.