Switch GH actions Rspec format to default progress
The "documentation" output is causing us to have to scroll through
thousands of output lines in order to find a failed test.

We are not getting any value of this detailed output at the moment, so
switching back to the more compact default output will help us to
quickly identify failed specs.