Set recaptcha failed attempts to Sentry as warning
We are currently registering these failed attempts to fill te recaptcha
as an exception/error.

This is not an exception we need to resolve, there is no point on
alerting about individual failures.

Downgrading it to warning will help us to have enough visibility, but
knowing is not an exception we need to act on. Alerts here should be
triggered based on hits/time. EG: 100 alerts in 1 hour. This would help
us to notice when a spam bot is attempting to fill our feedback forms.