Classify DFE-Sign-in errors as warnings for Sentry
An individual error on the Sign-in means nothing, it doesn't imply we
need to fix anything in our code/service, hence the warning level.

What we will do is define alerts on Sentry based on amount of errors per
time. So we will be able to know if the Sign-in service goes down.