Classify Vacancy import errors as Sentry warning
These alerts are for validation errors when importing vacancies from
external sources.

Having some vacancies failing to be imported due to missing info from
the source is a regular thing that doesn't need immediate action in our
side. We want to register it, be able to inspect it, etc, but an
occurrence of this is not an exception that needs to be raised in our
devs channel.

We are re-classifying these errors as "warnings", and we will create
alerts based on occurrences per time. So we will be able to notice if a
big amount of imports fail during an import.