Fix issue when submitting empty additional doc.
The validation bypass on the form was causing an exception on the
service, as users where selecting "yes" to upload an additional document
question, and then submitting the form without any file attached.

This bypass was introduced to force flow control method
"all_steps_valid?" being successfull so the flow can move to the next
step after choosing "no" when enquired about wanting to add additional
documents, or to allow to publish the job vacancy draft.

We are removing the bypass from the form, so the form itself will be
invalid unless there is a file. This prevents the user from submitting
an empty form when there is other files already attached to the vacancy.

At the same time, we are moving the bypass to the Vacancy form sequence,
so the form validity itself is not checked. Instead we ignore this step
when the vacancy has documents. A form without a document is invalid,
but "all_steps_valid?" check still returns true.