Fix issue when submitting empty additional doc.
The validation bypass on the form was causing an exception on the
service, as users where selecting "yes" to upload an additional document
question, and then submitting the form without any file attached.

The original attempt to fix this removed this validation bypass (when
there were other documents attached to the vacancy). But this caused
problems on the "Do you want to add another document? NO" flow, as it
relied into the next flow steps being all valid in order to send the
user to the final job draft/summary page. Since a form without an
attached document is not valid anymore, it was sending the user to the
document upload page even when

To avoid that I didn't found any other option but to hardcode the
redirection from the documents controller instead using the generic
"redirect to next step" method. Any other approach I have tried broke a
different flow.