Change redirection after deleting last document
When a publisher is editing a vacancy with additional documents, and
from the documents page it deletes all the documents, the service was
redirecting the publisher to the "upload a document" page.

Product decided this should redirect to the "Do you want to add an
aditional document?" page.

This will allow the publisher to select "No" and finish the flow or to
select "Yes" and then go to the upload document page, instead of
defaulting to the later.