Resolve through renaming documents form attribute
If we align the form document attribute name with the Vacancy documents
attribute name ("supporting_documents"), the "all_steps_valid?" check
will be able to automatically initialise with documents form passing in
the existing vacancy supporting documents into the form. The step will
be valid in this check.

This approach allows us to align the documents form with the rest and
removes the need of bypassing the step for validations.