Temporarily Increase Worker Count for BigQuery Backfill Operation
This commit increases the number of workers from 4 to 8 in our production
environment.  The purpose of this increase is to expedite a Rake task
that backfills entities from our production database to Big Query.

This is a temporary change, and the worker count will be reverted to 4
after the backfill operation has been completed.