Refactor: More obvious logic to display banner
The logic to display/hide the banner and its naming wasn't very clear.
Renamed it and adapted the code using it. Now the only purpose of the
method & why is overriden in the cookies preferences controller is
clearer.

Also stop using stubs for the helper in the System spec testing the
cookies banner display. Through a end-to-end spec using browser cookies
we are able to be sure that the banner display logic is correct.