Fix error on support dashboard date navigation
When the use cleared the date from the calendar datepicker in the
support dashboard, if then they click on "Go" the system would raise an
exception when trying to parse an empty date.

To resolve it, we use the default dates if the date selection is
submitted but contains an empty date.