Add "Quick Apply" tag to vacancy similar jobs
When a jobseeker is visiting a search listing, on the bottom of the
listing under the map there is a list of jobs similar to the current
vacancy the jobseeker is viewing.

For those similar jobs that allow to apply through Teaching Vacancies
service, we will display a "Quick Apply" tag under the link.