Move back to a single preference cookies selection
After considering the implications of splitting the cookies preference
settings into two different consents, the team decided to go back to a
single consent option.

We keep the changes informing the user about the collection of analytic
and marketing cookies. We also keep the change that will cause the
cookies banner to show up again for any existing user. Users will be
able to either accept/reject the additional cookies from the banner, or
be sent to the view the redesigned cookies preferences page.

From the cookies preferences page they would need to either accept or
reject the additional cookies.