Fix factorybot exceptions in production
Surprisingly our service is invoking Factorybot in production
environment for generate job application samples displayed to the users.

As production is multi-threaded, the use of Sequence calls in FactoryBot
causes thread exceptions when used in production.

The production code uses a particular factory trait, so I am modifying
it to avoid sequencing the job titles in that factory initialisation.

This way, we still have the factories able to use sequence calls in test
environments, while the production job application sample is safe as
doesn't sequence over the sample data.