Refactor Satisfaction Rating summaries
The code was querying data from the view instead of querying the data
and building the view with it. This proved problematic once we added a
second view (CSV in this case).

This refactor moves the querying to the controller and provides the
the data to the different views for presentation.