Explore generating CSV as views
This is a proof of concept, not complete and not meant to be merged, but
to discuss a different approach.

Given that the code to generate CSVs depends on helpers, and
conceptually we could take the CSV as an "output format" (as a json
formatter or xml formatter could be), I've explored building the CSV as
a view for a different format.

- It doesn't require external dependencies/gems.
- Doesn't add more code in the controller.
- Contains each CSV attributes logic in its respective view template,
  that loads the data from the controller as an HTML template would do.