Mitigate Support Users sign-in email issue
When trying to sign-in as a support user in our review apps, the
DFE-Sign-in is disabled and the login is done through one-off login
links sent by email.

Since our recent migration to Microsoft email hosting, Microsoft
introduced a certificate checker that proxies the sign-in links, and
results in the links becoming invalid when clicked as the proxy does a
double render that consumes the one-off link prior to redirect the user
to it.

This is causing a lot of headaches on our review environments, and as
the link is proxied, trying to find out the real link wrapped in the
proxy URL its nightmarish and not realistic for non-devs.

The solution here is to directly show the link as text, as done on the
publishers one-off login links. If you directly click the link it will
go through the proxy and fail, but at least here you will be able to see
the whole URL as a link, copy the text and paste it on your browser,
achieving the login.

This is not a proper fix to the root issue, but as we are still waiting
for support on that, lets make our team life less difficult meanwhile.