Improve Search of Similar jobs performance
The original code was retrieving all the search results in APP memory from the
DB. And then excluding and mapping them to get the ids.

We can improve the performance and reduce the memory allocation by
directly excluding and getting the IDs in the DB query. So only the IDs
get recorded in the APP memory and there is no need to loop through
them.