Memoise wider search suggestions (AKA Search killer)
This area is the sole response of all our 500ms-1s-infinite Vacancy
Search veeery slow queries.

It gets triggered when no vacancies were found and the search was done
over some location.

The whole idea of this is already bad enough (it is doing an expensive
search query for each one of the radius options!). But, on top of that,
the results weren't memoised and the whow thing with all the queries was
computed twice every time the Wider search suggestions were needed.

Memoising both the computing itself and the suggestions builder results
in the VacancySearch service.