Add missing not null DB restrictions
Tons of migrations adding not null restrictions to attributes.

Mainly needed to match rails ActiveRecord "belongs_to" associations.

Fix test

A build stubbed was being recorded in DB without ID because the lack of
NOT NULL restriction in the DB column.

build_stubbed is meant to fail when recorded in DB.