Add incremental backoff blocking to Rack::Attack
Throttling the connections per 4 seconds was enough for normal use
cases, but some uncontrolled spider bot is triggering this multiple
times withing a second.

These changes enable an incremental backoff blocking. When it increases
a higher limit will be blocked for the rest of the period during the
rule, and incrementally per time period rule.

The allowed requests per period and requests per second on each period
are quite generous.