Report a single Sentry error per source import
With the increase of number of external vacancy sources to import from,
and the increased rate of errors from some of the sources, reporting a
new transaction/error to Sentry for each of the failed vacancies on
every source and every import run is quickly consuming our Sentry
transaction quota.

Even if we increased the quota, this approach does not escale.

Instead, we are going to report a single transaction/error to Sentry,
containing data like:
- The total number of vacancies listed in the feed.
- The total number of vacancies that weren't valid/failed.
- The percentage of failed vacancies over the total.
- A list of all the failed external source references, together with
  their validation errors, so we can check the original feed (or the
  ImportedFailedVacancy associated object in our DB) to debug the issue.

This way, we reduce the number of transactions sent to sentry from
hundreds/thousands per hour to 4,5,6.. (the number of different source
providers) per hour.