Fix flaky test due to assumed order
This test is intermitently failing on CI runs.
We are ensuring the first vacancy is really the first vacancy by
ensuring the expiration date is clearly different from the other
vacancies instead of relying on a collection creation and picking up the
first from the collection of objects with milisecond differences.