Add given value on enumerable validation errors
At the moment when a value doesn't belong to the enumerable list, it
gets discarded when assigned and the validation errors turn into a
presence validation error that doesn't specify what value was attempted
to be registered.

Doing a check previous to the attempt to assign the value allows us to
record a specific validation error that will help us to debug failed
imports.