Don't try to save vacancy if contains enum errors
When we are capturing ArgumentErrors from the assign_attributes call,
these errors are being wiped when calling save over the vacancy, as it
re-triggers the validation and removes any existing custom error.

Instead, we want to keep those errors in the import failed vacancy, as
they provide better debugging information with the specific invalid
value provided to the enumerable field.