Fix: Do not override vacancy source parsing errors
When the source parsing adds an error to a Vacancy, this custom error
was getting overridden with the blind attempt to save the vacancy from
the orcherstrator class in a later stage.

The attempt to save triggers its own validation that erase any existing
error message attached to the vacancy object.

Example:
A vacancy containing an invalid job "FooBar" would get an error attached
as "base: ['FooBar is not a valid job_roles']" when trying to assign the
import source attributes to a new Vacancy object.

Then, when the field itself would become empty due to the failed
assignation: "vacancy.job_roles -> []"

If we try to do "vacancy.save" , it would faild validation but telling
that job_roles and any other following field in the assignation are
Empty.

So the consequences are:

1 - We get a "cannot be blank" error for a field that wasn't blank but
  contained wrong information. Missing essential information for its
  debugging.
2 - WE get a bunch of other field errors for following fields that were
  present but didn't get assigned due to the first error. These are not
  real errors and only add noise and confusion.

Solution:
If a Vacancy comes from the parsing stage with existing errors, do not
attempt to save it and just create a FailedImportedVacancy object with
the incoming error. This error will highlight the issue with the source
data.