Group Dependabot Bundler and NPM updates
Every week we get flooded with PRs from Dependabot.
For every PR, we need to run the whole deploy pipeline, and it takes
long time to go through them. It is taking a lot of time to catch up
with the PRs, and some gems are constantly getting updates.

This turned to be very disruptive towards our release pace, and 99% of
the times we just approve and merge the PR.

So what we will do here is for both gem and npm package updates:
- Excluding some specific libraries (govuk ones, packages that tend to
  break in updates...) open a single PR for all the gem updates by
  Dependabot that upgrade a minor/patch version. Same for the npm
  packages.
- Major versions or excluded packages for grouping will get their own
  PR.

This way we expect to drastically reduce the amount of PRs opened by
Dependabot and the deploy queues.