Create new Scheduled Maintenance Banner component
This banner is similar to the environment banner component, but it
won't have gradient background.

It will have a static standard GovUK blue colour and will be
configurable through instantiation of the date, start and end times for
the maintenance window.

Also, this banner will only be displayed in Production environment.