Upgrade documentation to use AKS instead of PaaS
We are upgrading the project documentation markdown files to present
information for our new Azure/Kubernetes hosting/tooling instead of the
removed GovUK PaaS/CloudFoundry info.

In some points, the way to proceed/reproduce the flow and instructions
with the new hosting is still unclear. Marked them as updated until we
learn how to proceed in AKS.