Update Redis Queue and Cache urls
After migrating to AKS from PaaS and upgrading the Sidekiq gem, we're
noticing an error with the parsing of the Queue and Cache Redis urls.

We're removing the extra values appened the both urls as they're valid
anymore.

TODO: we need to confirm with infrastructure if the solution in this
commit is valid or not.