Add landing page to account confirmation
Avoids the confirmation token being directly consumed when following the
email account confirmation link.

Users will need to click on a "Confirm" button to consume the one time
token and confirm the account.

This resolves an issue where Microsoft Cloud Security software is
pre-visiting the links and then redirecting the user to the safe link.
The first visit by the software consumes the one-time token, and the
user gets redirected to a "link expired" page.

By adding a landing page prior to consuming the token, the cloud
security check won't consume the token before the user lands.