Add "Already confirmed" page
Instead of relying on the "Link has expired" page with the resend button
for any confirmation error by Devise, we guard the confirmation page
from being reached if the user is already confirmed.

To do so, we will load a specific "Already confirmed" end page when the
already confirmed user follows the link for a second time.

This will help users experience, as multiple users have reported
confusion when landing on the Link has expired page and attempting to
resend an email that never arrives as they're already confirmed.