Fix format on plain text external job adverts
Some adverts from Broadbean and Every vacancy sources come as plain
text, without any HTML paragraph or line break code, but text return
characters.

That is causing them to be displayed as a single line, breaking their
formatting.

We are using simple_format to add the appropriate paragraph and line
text HTML tags on the display of those adverts.