Fix resend confirmation when email is not available
If the user tries to sign-in on an uncomfirmed account while their
session was previously clean from any user, Devise doesn't know where to
send the reconfirmation email to.

Attempting to follow the reconfirmation process was causing the users to
get a 404 page.

We have resolved this by identifying this case and requesting the
introduction of the email to send the reconfirmation to.

The email will only be sent if it belongs to a real unconfirmed user,
but we won't let the users know if they have introduced right/wrong
emails to avoid disclosing service emails information to third parties
testing email addresses through this form.