Fix Job Application seeds
The code generating the job applications was creating, for each vacancy,
an application for each one of the seeded Jobseekers. Each application
has one of the possible different statuses.

The problem is that this only worked as long as there are more listed
statuses than Jobseekers in the seeds. In the moment we had more
Jobseekers than statuses, the attempt to create an application for a new
jobseeker after consuming all the possible application statuses values
causes an exception when creating the factory.

Resolved it by reusing existing random status values once we have ran
over the whole values list and are still jobseekers left to create
applications for.