Move away from Rails Secrets file
The use of Rails secrets file got deprecated with Rails 5.2 in favour of
an encrypted credentias file with a master key.

In our service, the secrets file only contained the secret_key_base
value.

Moving to Rails 7.1, the usage of Rails Secrets file became incompatible
with the upgrade.

The good news is that, as in our service we use Environment Variables
stored on AWS Parameter Store instead of Secrets/Credentials, stop using
the secrets.yml file is as easy as removing the file.

"Rails.application.secret_key_base" method looks up for the value
defined as ENV variable first, then in the credentials and finally in
the secrets.

https://api.rubyonrails.org/classes/Rails/Application.html#method-i-secret_key_base

This allows us to don't have to change any infrastructure configuration,
as our non test/dev environments already had this value set up as an env variable.

Given we use env variables for configuring our services, we do not need
to generate a credentials file for the time being.