Update legacy migration task.
The Rails upgrade to version 7.1.2 is failing because an old migration
which changes data instead schema.

The error message suggests there is an Enum attribute ('contract_type') not
properly declared, but the content of
'ConvertNqtJobRoleToSuitableForNqtOnVacancies' migration lacks
the 'contract_type' attribute.

It appears it's an unexpected issue with Rails 7 and ActiveRecord enums -
when a migration uses a model which has an enum defined, ActiveRecord tries
to infer the type of the enum column in the database.
If the corresponding attribute is not in the table at the time
the migration is run, it leads to an error.