Add visa sponsorship details to searchable_content
Prior to this change, when a user enters the keyword visa, for example, the appropriate filter is selected but no results are shown. This is because
"visa sponsorship" is not shown within the searchable_content field so vacancies without the words visa or sponsorship in their role/description etc are filtered out. Adding "visa sponsorship" to searchable_content for vacancies which have visa_sponsorship_available: true fixes this issue