Allow Posts to have numbered headings
The markdown anchor ids removes any non-letter character from the start
of the id.
This caused the links for the posts not working when the post headings
are numbered.

We are resolving it by also removing the leading numbers from the
parameterized link id generated from the header in the view.
The links will point to the correct id for each header.