Only log Recaptcha errors when score is invalid
Instead of sending a recaptcha error to sentry every time that the
recaptcha error page is visited by a user, now the service will only
register an error when the recaptcha check failed.

This avoids hundreds of errors being sent per day to Sentry that weren't
real recaptcha failures but bots/users visiting the page url without
having gone through an error redirection.

Used this change as an opportunity to encapsulate all the recaptcha
controller logic into a controller concern, included only on the
controllers for the forms that need the recaptcha check.