Reduce DB updates from external source import jobs
Until now, the import from vacancy sources was re-writting in DB every
valid vacancy in each of the feeds, even if nothing was changed from the
already imported vacancy information.

Then, the updated_at data was used as a filter to remove any live
vacancy that wasn't updated in the current import job.

This causes a lot of DB writes on big feeds, and we have feeds with
almost 2k vacancies at the moment.

To avoid this, we only update the record in DB from the feed if either:
- It is a new record that wasn't persisted before.
- The attributes from the record have changed in the feed.

We will also keep a list of al the external reference identificators for
the feed vacancies, and use this as a criteria to search and mark as
removed any live vacancy previously imported and no longer present in
the feed.

We aim to drastically reduce the writes in DB with these changes.