Avoid unnecessary updates when importing vacancy without Subjects
Noticed that when the subjects optional field is coming null, when
building the vacancy update attributes over the already created vacancy
Rails thinks the subjects attribute changed and updates the Vacancy in
DB. This process is costly in feeds like MyNewTerm with thousands of
vacancies.

This happens because the "no vacancies" is stored as "[]" in DB and then
compared against the incoming "nil". And this happens for every run.