Normalise vacancy data before checking for changes
Debugging why so many MyNewTerm vacancies get updated over and over with
each hourly pull, causing stress on the DB due to hundreds/thousands of
writes being commanded in a short span of time, we noticed that for many
of those, the feed includes invalid values that get discarded with an
ActiveRecord callback that ensures data being normalised.

EG: Vacancies cannot have subjects unless the vacancy is belongs to a
particular phase.

So, when those vacancies are pulled and saved, gets written in DB with
some values discarded, then when we pull again, we populate the vacancy
attributes with the data from the pull. Check if it changed from the DB
original values, notices the disparity between:
- The attribute value resetted to 'nil' or "" in DB.
- The attribute populated from the feed.
And then proceeds to try to save the new value. The value gets discarded
on the AR callback prior to the #save...the vacancy gets updated in DB
without any actual change.

To avoid this, we apply the values normalisation prior to checking for
changes and attempting to save.