Remove leading&trailing spaces from MyNewTerm texts
These fields are stripped when stored in our DB, but the presence of
leading/trailing spaces in the feed makes the imported vacancy be
considered as changed from the information already stored in DB, causing
a further DB save.

Striping the leading and trailing spaces when parsing the fields avoids
this issue.