Fix Ark integration org association to vacancies
There is a vacancy from Ark that got associated with 1902 schools
because a bug on this bit.

The bug is caused by searching Orgs by URN even when the URN is nil,
what effectively associates the vacancy with any org in the system
without an URN (1902 at the moment).

Applying same Org/Trust associations rules as recently brought to other
integrations.