Deduplicate scopes on vacancy search query
The SQL query built by the VacancySearch service resulted into a SQL
query that contained repeated conditions:
- The "published" state is checked multiple times.
- The "publish_on" and "expires_at" values are checked multiple times.

This makes the logs and debugging harder to read, and even if the
performance difference is nelegible (pretty sure PostgreSQL optimices
this), is better to tidy this up to avoid repeated conditions in the
queries.

Also, having a scope definition (meant to be chained) relying on the
"live" scope is unnecessary and may induce to mistakes (I wouldn't
assume a Location/Filter scope also chains the "live" scope to the
query).