Avoid unnecessary table joins from count queries
For the Vacancy count method, relying on "size" instead of
"count(:id)" transforms the query into a simpler one, saving two "LEFT
OUTER JOIN" calls.

This is a weird one.

If we count methods as used to do originally prior to the sorting by
distance:
- vacancies.count
Then the formed query would raise a malformed SQL syntax error with the
sort by distance.

With the current implementation to make it compatible with the sort by
distance:
- vacancies.count(:id)
Then the count query weirdly forces LEFT OUTER JOIN relationships with
both "organisations" and "organisation_vacancies". This is odd as these
joins are triggered even when we are not filtering by any organisation
value. Thos overcomplicates the query, when it doesn't seem to be any
need for it.

If, instead, we rely on the "size" method:
- vacancies.size
Then the count query doesn't do any join with the organisation or
organisation_vacancies tables if we don't filter by org fields.
And if we filter by org fields, it does simple inner joins.