Re-enable Pagy original pagination mode
After DB performance issues caused service outages while running the
pagination in the default mode. We switched to countless mode, that
seemed to avoid these issues.

Now that we have reduced the amount of queries triggered per search, and
optimized the queries triggered by pagy, we want to re-enable the
default pagination as it used to work.