Exclude working_patterns_details from Vacancy copy
Many recent vacancies contain a value for "working_patterns_fields",
even when this value is not set by the user on any form.

The value is being populated by the VacancyCopy service, as the
publishers use it frequently to create new vacancies from older ones.

We need to exclude it as its value clashes with the actual ones
introduced by users ("full_time_details", "part_time_details").