Stop blocking forms based on Recaptcha v3 score
The Google Recaptcha V3 tends to give low scoring to lots of legit
requests when there is a surge of submissions coming from our marketing
campaigns.

This causes a lot of friction with users that get their form submissions
blocked without any recourse/explanation.

Google Recaptcha V3 is meant to give the service more power to do
further fine tuning checks based on the Recaptcha score. Instead, we are
using it as a PASS/FAIl test that block the users.

We are moving to log the low scores into Sentry, but not block the form
submission.