Fix Publisher email code backup sign-in flow
When DFE Sign-in service is not enabled, our service use login keys sent
by email as a backup authentication method for Publishers.

This is widely used on local development and review apps.

Since the Microsoft Outlook security check pre-visits any HTTPS link
received by email, and our service was consuming the login token when
the publisher follows the log-in email, the token was being consumed by
the Outlook security check prior to the Publisher landing on the page.
As a consequence, the publisher found an invalid page and was unable to
sign-in.

To resolve this, we redesigned how the publisher sign-in process work:
- When following the link, the token will be checked but not consumed
  (so the https security can be checked by Microsoft Outlook).
- Then, if the token is valid, the publisher will be shown a page to
  choose the organisation to sign-in for.
- Upon selecting the organisation and submitting the choice, the login
  token will be consumed/deleted and the publisher signed in for the
  choosen org.