Introduce Recaptcha v3 check with v2 fallback
Recaptcha v3 is used as a first check on the form submissions using
Recaptcha.
When the minimum score threshold is not achievend, instead of ignoring
it or blocking the user, we will display a Recaptcha v2 check on the
form next to the submission button, enquiring the user to complete it.

Then, the user may be requested to resolve a Recaptcha puzzle for being
able to submit the form.